test: run cargo-flux tests through xtask
