+++ crates/flux-bin/src/bin/cargo-flux.rs
@@ -1,22 +1,11 @@
-use std::{
-    self, env,
-    io::{BufWriter, Write},
-    process::{Command, exit},
-};
+use std::process::exit;
 
-use anyhow::anyhow;
-use cargo_metadata::{Metadata, camino::Utf8Path};
 use clap::Parser as _;
 use flux_bin::{
-    FluxMetadata,
+    cargo_flux,
     cargo_flux_opts::{CargoFluxCommand, Cli},
-    utils::{
-        EXIT_ERR, flux_sysroot_dir, get_binary_path, get_flux_driver_path, get_rust_toolchain,
-        print_version_and_exit,
-    },
+    utils::{EXIT_ERR, print_version_and_exit},
 };
-use itertools::Itertools;
-use tempfile::NamedTempFile;
 
 fn main() {
     let Cli::Flux { check_opts, command, version, verbose } = Cli::parse();
@@ -28,315 +17,11 @@ fn main() {
 
     let command = command.unwrap_or(CargoFluxCommand::Check(check_opts));
 
-    match run(command) {
+    match cargo_flux::run(command) {
         Ok(exit_code) => exit(exit_code),
         Err(e) => {
             println!("Failed to run `cargo-flux`, error={e}");
             exit(EXIT_ERR)
         }
-    };
-}
-
-fn run(cargo_flux_cmd: CargoFluxCommand) -> anyhow::Result<i32> {
-    let toolchain = get_rust_toolchain()?;
-    let cargo_path = get_binary_path(&toolchain, "cargo")?;
-
-    let metadata = cargo_flux_cmd.metadata().cargo_path(&cargo_path).exec()?;
-    let sysroot = flux_sysroot_dir();
-    let flux_driver_path = get_flux_driver_path(&sysroot)?;
-    let flux_flags = env_flags("FLUXFLAGS");
-    let local_flux_flags = local_env_flags();
-    let config_file =
-        write_cargo_config(metadata, &sysroot, &cargo_flux_cmd, &flux_flags, &local_flux_flags)?;
-
-    let mut cargo_command = cargo_command(&toolchain, &flux_driver_path);
-
-    cargo_flux_cmd.forward_args(&mut cargo_command, config_file.path());
-
-    Ok(cargo_command.status()?.code().unwrap_or(EXIT_ERR))
-}
-
-fn env_flags(name: &str) -> Vec<String> {
-    env::var(name)
-        .map(|flags| flags.split(" ").map(Into::into).collect())
-        .unwrap_or_default()
-}
-
-fn local_env_flags() -> Vec<String> {
-    env::var("FLUXLOCALFLAGS")
-        .map(|flags| parse_local_flags(&flags))
-        .unwrap_or_default()
-}
-
-fn parse_local_flags(flags: &str) -> Vec<String> {
-    flags.split_whitespace().map(Into::into).collect()
-}
-
-fn cargo_command(toolchain: &str, flux_driver_path: &std::path::Path) -> Command {
-    let mut command = Command::new("cargo");
-
-    // We set `RUSTC` as an environment variable and not in in the [build]
-    // section of the config file to make sure we run flux even when the
-    // variable is already set. We also unset `RUSTC_WRAPPER` to avoid
-    // conflicts, e.g., see https://github.com/flux-rs/flux/issues/1155.
-    // FLUXLOCALFLAGS is consumed by cargo-flux and must not leak into children.
-    command
-        .env("RUSTC", flux_driver_path)
-        .env("RUSTC_WRAPPER", "")
-        .env_remove("FLUXLOCALFLAGS")
-        .arg(format!("+{toolchain}"));
-    command
-}
-
-fn write_cargo_config(
-    metadata: Metadata,
-    sysroot: &std::path::Path,
-    cargo_flux_cmd: &CargoFluxCommand,
-    flux_flags: &[String],
-    local_flux_flags: &[String],
-) -> anyhow::Result<NamedTempFile> {
-    let flux_toml = config::Config::builder()
-        .add_source(config::File::with_name("flux.toml").required(false))
-        .build()?;
-
-    if flux_toml.get_bool("enabled").is_ok() {
-        return Err(anyhow!("`enabled` cannot be set in `flux.toml`"));
     }
-    let targeted_package_ids = cargo_flux_cmd.targeted_package_ids(&metadata);
-
-    let mut file = NamedTempFile::new()?;
-    {
-        let mut w = BufWriter::new(&mut file);
-        write!(
-            w,
-            r#"
-[unstable]
-profile-rustflags = true
-
-[env]
-FLUX_BUILD_SYSROOT = "1"
-FLUX_CARGO = "1"
-
-[profile.flux]
-inherits = "dev"
-incremental = false
-        "#
-        )?;
-
-        for package in metadata.packages {
-            let package_local_flags =
-                if targeted_package_ids.contains(&package.id) { local_flux_flags } else { &[] };
-            let flux_metadata: FluxMetadata = config::Config::builder()
-                .add_source(FluxMetadataSource::new(
-                    package.manifest_path.to_string(),
-                    package.metadata,
-                ))
-                .add_source(flux_toml.clone())
-                .build()?
-                .try_deserialize()?;
-
-            if flux_metadata.enabled {
-                // For workspace members, cargo sets the workspace's root as the working dir
-                // when running flux. Paths will be relative to that, so we must normalize
-                // glob patterns to be relative to the workspace's root.
-                let manifest_dir_relative_to_workspace = package
-                    .manifest_path
-                    .strip_prefix(&metadata.workspace_root)
-                    .ok()
-                    .and_then(Utf8Path::parent);
-
-                let sysroot_flag = format!("-Fsysroot={}", sysroot.display());
-                write!(
-                    w,
-                    r#"
-[profile.flux.package."{}"]
-rustflags = [{:?}]
-                        "#,
-                    package.id,
-                    flux_metadata
-                        .into_flags(
-                            &metadata.target_directory,
-                            manifest_dir_relative_to_workspace,
-                            cargo_flux_cmd
-                                .only_check()
-                                .filter(|_| targeted_package_ids.contains(&package.id))
-                        )
-                        .iter()
-                        .chain(flux_flags)
-                        .chain(package_local_flags)
-                        .map(|s| s.as_str())
-                        .chain(["-Fverify=on", "-Ffull-compilation=on", sysroot_flag.as_str()])
-                        .format(", ")
-                )?;
-            }
-        }
-    }
-    Ok(file)
-}
-
-#[derive(Clone, Debug)]
-struct FluxMetadataSource {
-    origin: String,
-    value: serde_json::Value,
-}
-
-impl FluxMetadataSource {
-    fn new(origin: String, value: serde_json::Value) -> Self {
-        Self { origin, value }
-    }
-}
-
-impl config::Source for FluxMetadataSource {
-    fn clone_into_box(&self) -> Box<dyn config::Source + Send + Sync> {
-        Box::new(self.clone())
-    }
-
-    fn collect(&self) -> Result<config::Map<String, config::Value>, config::ConfigError> {
-        if let serde_json::Value::Object(metadata) = &self.value
-            && let Some(flux_metadata) = metadata.get("flux")
-        {
-            let config_value = serde_json_to_config(flux_metadata, &self.origin)?;
-            if let config::ValueKind::Table(table) = config_value.kind {
-                Ok(table)
-            } else {
-                Err(config::ConfigError::Message("expected a table".to_string()))
-            }
-        } else {
-            Ok(Default::default())
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::fs;
-
-    use super::*;
-
-    fn generated_config(args: &[&str], global_flags: &[&str], local_flags: &[&str]) -> String {
-        let Cli::Flux { check_opts, command, .. } = Cli::parse_from(args);
-        let command = command.unwrap_or(CargoFluxCommand::Check(check_opts));
-        let metadata = command.metadata().exec().unwrap();
-        let global_flags = global_flags
-            .iter()
-            .map(|flag| flag.to_string())
-            .collect::<Vec<_>>();
-        let local_flags = local_flags
-            .iter()
-            .map(|flag| flag.to_string())
-            .collect::<Vec<_>>();
-        let config = write_cargo_config(
-            metadata,
-            &flux_sysroot_dir(),
-            &command,
-            &global_flags,
-            &local_flags,
-        )
-        .unwrap();
-        fs::read_to_string(config.path()).unwrap()
-    }
-
-    fn package_profile<'a>(config: &'a str, package: &str) -> &'a str {
-        config
-            .split("[profile.flux.package.")
-            .find(|profile| profile.contains(package))
-            .unwrap_or_else(|| panic!("generated profile for package {package}"))
-    }
-
-    #[test]
-    fn local_flags_are_only_added_to_targeted_packages() {
-        let global_flags = &["-Fcache=off"];
-        let local_flags = &["-Fsummary=off", "-Fdump-constraint=on"];
-        let config = generated_config(
-            &["cargo", "flux", "check", "-p", "flux-alloc"],
-            global_flags,
-            local_flags,
-        );
-        let targeted = package_profile(&config, "flux-alloc");
-        let dependency = package_profile(&config, "flux-core");
-
-        assert!(targeted.contains("-Fcache=off"));
-        assert!(targeted.contains("-Fsummary=off"));
-        assert!(targeted.contains("-Fdump-constraint=on"));
-        assert!(dependency.contains("-Fcache=off"));
-        assert!(!dependency.contains("-Fsummary=off"));
-        assert!(!dependency.contains("-Fdump-constraint=on"));
-
-        let config = generated_config(
-            &["cargo", "flux", "check", "-p", "flux-alloc", "-p", "flux-core"],
-            global_flags,
-            local_flags,
-        );
-        assert!(package_profile(&config, "flux-alloc").contains("-Fsummary=off"));
-        assert!(package_profile(&config, "flux-core").contains("-Fsummary=off"));
-
-        let config = generated_config(
-            &["cargo", "flux", "check", "--workspace", "--exclude", "flux-core"],
-            global_flags,
-            local_flags,
-        );
-        assert!(package_profile(&config, "flux-alloc").contains("-Fsummary=off"));
-        assert!(!package_profile(&config, "flux-core").contains("-Fsummary=off"));
-    }
-
-    #[test]
-    fn local_flags_ignore_surrounding_and_repeated_whitespace() {
-        assert_eq!(
-            parse_local_flags("  -Fsummary=off\t\n-Fdump-constraint=on  "),
-            ["-Fsummary=off", "-Fdump-constraint=on"]
-        );
-    }
-
-    #[test]
-    fn local_flags_are_removed_from_the_cargo_environment() {
-        let command = cargo_command("nightly", std::path::Path::new("flux-driver"));
-
-        assert!(
-            command
-                .get_envs()
-                .any(|(name, value)| { name == "FLUXLOCALFLAGS" && value.is_none() })
-        );
-    }
-}
-
-fn serde_json_to_config(
-    value: &serde_json::Value,
-    origin: &String,
-) -> Result<config::Value, config::ConfigError> {
-    let kind = match value {
-        serde_json::Value::Null => config::ValueKind::Nil,
-        serde_json::Value::Bool(b) => config::ValueKind::Boolean(*b),
-        serde_json::Value::Number(number) => {
-            if let Some(n) = number.as_u128() {
-                config::ValueKind::U128(n)
-            } else if let Some(n) = number.as_i128() {
-                config::ValueKind::I128(n)
-            } else if let Some(n) = number.as_u64() {
-                config::ValueKind::U64(n)
-            } else if let Some(n) = number.as_i64() {
-                config::ValueKind::I64(n)
-            } else if let Some(n) = number.as_f64() {
-                config::ValueKind::Float(n)
-            } else {
-                return Err(config::ConfigError::Message("invalid number".to_string()));
-            }
-        }
-        serde_json::Value::String(s) => config::ValueKind::String(s.clone()),
-        serde_json::Value::Array(values) => {
-            config::ValueKind::Array(
-                values
-                    .iter()
-                    .map(|v| serde_json_to_config(v, origin))
-                    .try_collect()?,
-            )
-        }
-        serde_json::Value::Object(map) => {
-            config::ValueKind::Table(
-                map.iter()
-                    .map(|(k, v)| Ok((k.clone(), serde_json_to_config(v, origin)?)))
-                    .try_collect()?,
-            )
-        }
-    };
-    Ok(config::Value::new(Some(origin), kind))
 }
